LPS-55530 Move CMIS classes to its module, removing usages
