+++ portal-impl/src/com/liferay/portal/convert/ConvertDocumentLibrary.java
@@ -42,7 +42,6 @@ import com.liferay.portlet.documentlibrary.model.DLFileEntry;
 import com.liferay.portlet.documentlibrary.model.DLFolderConstants;
 import com.liferay.portlet.documentlibrary.service.DLFileEntryLocalServiceUtil;
 import com.liferay.portlet.documentlibrary.store.AdvancedFileSystemStore;
-import com.liferay.portlet.documentlibrary.store.CMISStore;
 import com.liferay.portlet.documentlibrary.store.DBStore;
 import com.liferay.portlet.documentlibrary.store.FileSystemStore;
 import com.liferay.portlet.documentlibrary.store.JCRStore;
@@ -399,9 +398,8 @@ public class ConvertDocumentLibrary
 	private static final String _FILE_SYSTEM_STORE_SUFFIX = "FileSystemStore";
 
 	private static final String[] _HOOKS = new String[] {
-		AdvancedFileSystemStore.class.getName(), CMISStore.class.getName(),
-		DBStore.class.getName(), FileSystemStore.class.getName(),
-		JCRStore.class.getName()
+		AdvancedFileSystemStore.class.getName(), DBStore.class.getName(),
+		FileSystemStore.class.getName(), JCRStore.class.getName()
 	};
 
 	private static final Log _log = LogFactoryUtil.getLog(
