+++ portal-impl/src/com/liferay/portlet/documentlibrary/store/StoreFactory.java
@@ -170,10 +170,6 @@ public class StoreFactory {
 			AdvancedFileSystemStore.class.getName()
 		},
 		new String[] {
-			"com.liferay.documentlibrary.util.CMISHook",
-			CMISStore.class.getName()
-		},
-		new String[] {
 			"com.liferay.documentlibrary.util.FileSystemHook",
 			FileSystemStore.class.getName()
 		},
