+++ portal-impl/src/portal.properties
@@ -8984,7 +8984,6 @@
     # document library server will use this to persist documents.
     #
     #dl.store.impl=com.liferay.portlet.documentlibrary.store.AdvancedFileSystemStore
-    #dl.store.impl=com.liferay.portlet.documentlibrary.store.CMISStore
     #dl.store.impl=com.liferay.portlet.documentlibrary.store.DBStore
     dl.store.impl=com.liferay.portlet.documentlibrary.store.FileSystemStore
     #dl.store.impl=com.liferay.portlet.documentlibrary.store.JCRStore
